feat(python) * equity codility
