+++ eqt_codility/threeDigits.py
