+++ eqt_codility/time_series.py
