+++ eqt_codility/turing.py
